# openapi.yaml
openapi: 3.1.0
info:
  title: Astradio API
  version: "1.0"
servers:
  - url: https://api.astradio.io/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /compare/{username}:
    post:
      summary: Overlay my chart with another user if allowed
      parameters: [{ name: username, in: path, required: true, schema: {type: string} }]
      responses: { "200": { description: Overlay track queued or returned } }